Move FolderType and FolderClass out of Folder
