+++ app/core/src/main/java/com/fsck/k9/mailstore/BackendStorageListener.kt
@@ -1,7 +1,7 @@
 package com.fsck.k9.mailstore
 
 import com.fsck.k9.backend.api.FolderInfo
-import com.fsck.k9.mail.Folder.FolderType
+import com.fsck.k9.mail.FolderType
 
 interface BackendStorageListener {
     fun onFoldersCreated(folders: List<FolderInfo>)
