+++ app/core/src/main/java/com/fsck/k9/mailstore/FolderTypeConverter.kt
@@ -1,7 +1,7 @@
 @file:JvmName("FolderTypeConverter")
 package com.fsck.k9.mailstore
 
-import com.fsck.k9.mail.Folder.FolderType
+import com.fsck.k9.mail.FolderType
 
 @JvmName("fromDatabaseFolderType")
 fun String.toFolderType(): FolderType {
