+++ app/core/src/main/java/com/fsck/k9/mailstore/LocalFolder.java
@@ -21,8 +21,8 @@ import com.fsck.k9.mail.BodyPart;
 import com.fsck.k9.mail.BoundaryGenerator;
 import com.fsck.k9.mail.FetchProfile;
 import com.fsck.k9.mail.Flag;
-import com.fsck.k9.mail.Folder.FolderClass;
-import com.fsck.k9.mail.Folder.FolderType;
+import com.fsck.k9.mail.FolderClass;
+import com.fsck.k9.mail.FolderType;
 import com.fsck.k9.mail.Message;
 import com.fsck.k9.mail.Message.RecipientType;
 import com.fsck.k9.mail.MessageRetrievalListener;
