+++ app/core/src/main/java/com/fsck/k9/search/AccountSearchConditions.kt
@@ -3,7 +3,7 @@ package com.fsck.k9.search
 import com.fsck.k9.Account
 import com.fsck.k9.Account.FolderMode
 import com.fsck.k9.BaseAccount
-import com.fsck.k9.mail.Folder.FolderClass
+import com.fsck.k9.mail.FolderClass
 import com.fsck.k9.search.SearchSpecification.Attribute
 import com.fsck.k9.search.SearchSpecification.SearchCondition
 import com.fsck.k9.search.SearchSpecification.SearchField
