+++ app/core/src/test/java/com/fsck/k9/mailstore/K9BackendFolderTest.kt
@@ -10,7 +10,7 @@ import com.fsck.k9.backend.api.BackendFolder
 import com.fsck.k9.backend.api.FolderInfo
 import com.fsck.k9.mail.Address
 import com.fsck.k9.mail.Flag
-import com.fsck.k9.mail.Folder.FolderType
+import com.fsck.k9.mail.FolderType
 import com.fsck.k9.mail.internet.MimeMessage
 import com.fsck.k9.mail.internet.MimeMessageHelper
 import com.fsck.k9.mail.internet.TextBody
