+++ app/ui/src/main/java/com/fsck/k9/activity/setup/FolderSettings.java
@@ -19,7 +19,7 @@ import com.fsck.k9.activity.K9PreferenceActivity;
 import com.fsck.k9.controller.MessagingController;
 import com.fsck.k9.job.K9JobManager;
 import com.fsck.k9.mail.Folder;
-import com.fsck.k9.mail.Folder.FolderClass;
+import com.fsck.k9.mail.FolderClass;
 import com.fsck.k9.mail.MessagingException;
 import com.fsck.k9.mailstore.LocalFolder;
 import com.fsck.k9.mailstore.LocalStore;
