+++ app/ui/src/main/java/com/fsck/k9/ui/folders/FolderIconProvider.kt
@@ -2,7 +2,7 @@ package com.fsck.k9.ui.folders
 
 import android.content.res.Resources
 import android.util.TypedValue
-import com.fsck.k9.mail.Folder.FolderType as LegacyFolderType
+import com.fsck.k9.mail.FolderType as LegacyFolderType
 import com.fsck.k9.mailstore.FolderType
 import com.fsck.k9.ui.R
