+++ backend/api/src/main/java/com/fsck/k9/backend/api/FolderInfo.kt
@@ -1,5 +1,5 @@
 package com.fsck.k9.backend.api
 
-import com.fsck.k9.mail.Folder
+import com.fsck.k9.mail.FolderType
 
-data class FolderInfo(val serverId: String, val name: String, val type: Folder.FolderType)
+data class FolderInfo(val serverId: String, val name: String, val type: FolderType)
