+++ backend/pop3/src/main/java/com/fsck/k9/backend/pop3/CommandRefreshFolderList.kt
@@ -2,7 +2,7 @@ package com.fsck.k9.backend.pop3
 
 import com.fsck.k9.backend.api.BackendStorage
 import com.fsck.k9.backend.api.FolderInfo
-import com.fsck.k9.mail.Folder.FolderType
+import com.fsck.k9.mail.FolderType
 import com.fsck.k9.mail.store.pop3.Pop3Folder
 
 internal class CommandRefreshFolderList(private val backendStorage: BackendStorage) {
