+++ mail/common/src/main/java/com/fsck/k9/mail/FolderClass.java
@@ -1,0 +1,6 @@
+package com.fsck.k9.mail;
+
+// NONE is obsolete, it will be translated to NO_CLASS for display and to INHERITED for sync and push
+public enum FolderClass {
+    NONE, NO_CLASS, INHERITED, FIRST_CLASS, SECOND_CLASS
+}
