+++ mail/common/src/main/java/com/fsck/k9/mail/FolderType.java
@@ -1,0 +1,12 @@
+package com.fsck.k9.mail;
+
+public enum FolderType {
+    REGULAR,
+    INBOX,
+    OUTBOX,
+    DRAFTS,
+    SENT,
+    TRASH,
+    SPAM,
+    ARCHIVE
+}
