+++ mail/protocols/imap/src/main/java/com/fsck/k9/mail/store/imap/FolderListItem.kt
@@ -1,5 +1,5 @@
 package com.fsck.k9.mail.store.imap
 
-import com.fsck.k9.mail.Folder.FolderType
+import com.fsck.k9.mail.FolderType
 
 data class FolderListItem(val name: String, val type: FolderType)
